WIP: Work on trying to get things working with patch endpoint.
