+++ packages/cds-core/src/cds_core/remote.py
@@ -1,4 +1,5 @@
 import hashlib
+import json
 import os
 from functools import cached_property
 
@@ -9,12 +10,16 @@ from solara_enterprise import auth
 
 from .base_states import BaseAppState, BaseStoryState, BaseStageState, BaseState
 from .logger import setup_logger
+from .utils import CDSJSONEncoder
 
 logger = setup_logger("API")
 
 
 class BaseAPI:
-    API_URL = "https://api.cosmicds.cfa.harvard.edu"
+    # API_URL = "https://api.cosmicds.cfa.harvard.edu"
+    API_URL = "http://localhost:8081"
+
+    initial_load_completed = False
 
     @cached_property
     def request_session(self):
@@ -208,6 +213,9 @@ class BaseAPI:
                 .get("state", None)
             )
 
+            logger.info("Story JSON")
+            logger.info(story_json)
+
             if story_json is None:
                 logger.error(
                     f"Failed to retrieve state for story {local_state.value.story_id} "
@@ -235,6 +243,11 @@ class BaseAPI:
 
         logger.info("Updated state from database.")
 
+        logger.info(global_state.value)
+        logger.info(local_state.value)
+
+        self.initial_load_completed = True
+
         return local_state.value
 
     def put_story_state(
@@ -243,6 +256,35 @@ class BaseAPI:
         local_state: Reactive[BaseStoryState],
     ):
         raise NotImplementedError()
+
+    def patch_story_state(
+        self, patch: dict, global_state: Reactive[BaseAppState], local_state: Reactive[BaseStoryState]
+    ):
+        if not global_state.value.update_db or self.is_educator:
+            logger.info("Skipping DB write")
+            return False
+
+        logger.info("Serializing state into DB.")
+
+        state = {
+            "app": patch,
+        }
+
+        state_json = json.dumps(state, cls=CDSJSONEncoder)
+        logger.info("Payload for patch request")
+        logger.info(state_json)
+        r = self.request_session.patch(
+            f"{self.API_URL}/story-state/{global_state.value.student.id}/{local_state.value.story_id}",
+            headers={"Content-Type": "application/json"},
+            data=state_json,
+        )
+
+        if r.status_code != 200:
+            logger.error("Failed to write story state to database.")
+            logger.error(r.text)
+            return False
+
+        return True
 
     @staticmethod
     def clear_user(state: Reactive[BaseAppState]):
